requirements.txt
alembic==1.8.1
aniso8601==9.0.1
autopep8==2.0.0
bcrypt==4.0.1
click==8.1.3
ecdsa==0.18.0
Flask==2.2.2
Flask-RESTful==0.3.9
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
python-jose==3.3.0
pytz==2022.6
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
